requirements.txt
backcall==0.1.0
backports.csv==1.0.7
certifi==2019.9.11
chardet==3.0.4
chart-studio==1.0.0
colorama==0.4.1
colorlover==0.3.0
decorator==4.4.1
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==2.2.2
django-filter==2.2.0
django-heroku==0.3.1
django-import-export==1.2.0
django-staticfiles-moment==2.0.0
et-xmlfile==1.0.1
gunicorn==19.9.0
idna==2.8
ipython==7.9.0
ipython-genutils==0.2.0
jdcal==1.4.1
jedi==0.15.1
mysql-connector-python==8.0.16
numpy==1.17.2
odfpy==1.4.0
openpyxl==3.0.0
pandas==0.25.1
parso==0.5.1
pickleshare==0.7.5
playsound==1.2.2
plotly==4.3.0
prompt-toolkit==2.0.10
protobuf==3.8.0
psycopg2==2.8.2
Pygments==2.4.2
python-dateutil==2.8.0
python-decouple==3.1
pytz==2019.1
PyYAML==5.1.2
requests==2.22.0
retrying==1.3.3
six==1.12.0
sqlparse==0.3.0
tablib==0.13.0
traitlets==4.3.3
urllib3==1.25.7
virtualenv==16.6.0
wcwidth==0.1.7
whitenoise==4.1.2
xlrd==1.2.0
xlwt==1.3.0
